feat: Refactor ValidatedTextField Handle isEdit without a ternary operator
